+++ asf_search/ASFSearchOptions/config.py
@@ -1,9 +1,8 @@
 from asf_search.constants import INTERNAL
 from asf_search.ASFSession import ASFSession
 
-defaults = {
-    'host': INTERNAL.SEARCH_API_HOST,
+config = {
+    'host': INTERNAL.CMR_HOST,
     'provider': INTERNAL.DEFAULT_PROVIDER,
     'session': ASFSession(),
-    'maturity': None
 }