+++ asf_search/search/__init__.py
@@ -4,3 +4,4 @@ from .product_search import product_search
 from .geo_search import geo_search
 from .baseline_search import stack_from_id, get_stack_opts
 from .campaigns import campaigns
+from .search_count import search_count